pyproject.toml
[tool.poetry]
name = "superbench"
version = "0.1.0"
description = ""
authors = ["Thomas Chiang <t7hiang@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "==3.13.5"
beautifulsoup4 = "*"
langchain-community = "*"
tiktoken = "*"
langchainhub = "*"
langchain = "*"
langgraph = "*"
langchain-openai = "*"
python-dotenv = "*"
black = "*"
isort = "*"
langchain-google-genai = "*"
pinecone = "*"
langchain-pinecone = "^0.2.8"
fastapi = {extras = ["standard"], version = "^0.115.13"}
firebase-admin = "^6.9.0"
google-auth-oauthlib = "^1.2.2"
google-api-python-client = "^2.172.0"
requests = "^2.32.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"